Solution to Calculate and Print the Area of a Circle

+++ Calculate and Print the Area of a Circle@@6788d5012db25efec946f943/code.c
@@ -1,8 +1,8 @@
 #include<stdio.h>
 int main(){
 
-    int rad;
-    scanf("%d",&rad);
+    float rad;
+    scanf("%f",&rad);
     float area = 3.14*rad*rad;
     print("%.2f",area);
     return 0;
